+++ fluent-syntax/test/util.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs from "fs";
 
 function nonBlank(line) {
   return !/^\s*$/.test(line);
@@ -11,12 +11,12 @@ function countIndent(line) {
 
 export function ftl(strings) {
   const [code] = strings;
-  const lines = code.split('\n').slice(1, -1);
+  const lines = code.split("\n").slice(1, -1);
   const indents = lines.filter(nonBlank).map(countIndent);
   const common = Math.min(...indents);
   const indent = new RegExp(`^\\s{${common}}`);
-  const dedented = lines.map(line => line.replace(indent, ''));
-  return `${dedented.join('\n')}\n`;
+  const dedented = lines.map(line => line.replace(indent, ""));
+  return `${dedented.join("\n")}\n`;
 }
 
 export function readfile(path) {