+++ fluent/src/parser.js
@@ -5,6 +5,9 @@ const MAX_PLACEABLES = 100;
 const entryIdentifierRe = /-?[a-zA-Z][a-zA-Z0-9_-]*/y;
 const identifierRe = /[a-zA-Z][a-zA-Z0-9_-]*/y;
 const functionIdentifierRe = /^[A-Z][A-Z_?-]*$/;
+const unicodeEscapeRe = /^[a-fA-F0-9]{4}$/;
+const trailingWSRe = /[ \t\n\r]+$/;
+
 
 /**
  * The `Parser` class is responsible for parsing FTL resources.
@@ -74,44 +77,13 @@ class RuntimeParser {
     const ch = this._source[this._index];
 
     // We don't care about comments or sections at runtime
-    if (ch === "/" ||
-      (ch === "#" &&
-        [" ", "#", "\n"].includes(this._source[this._index + 1]))) {
+    if (ch === "#" &&
+        [" ", "#", "\n"].includes(this._source[this._index + 1])) {
       this.skipComment();
       return;
     }
 
-    if (ch === "[") {
-      this.skipSection();
-      return;
-    }
-
     this.getMessage();
-  }
-
-  /**
-   * Skip the section entry from the current index.
-   *
-   * @private
-   */
-  skipSection() {
-    this._index += 1;
-    if (this._source[this._index] !== "[") {
-      throw this.error('Expected "[[" to open a section');
-    }
-
-    this._index += 1;
-
-    this.skipInlineWS();
-    this.getVariantName();
-    this.skipInlineWS();
-
-    if (this._source[this._index] !== "]" ||
-        this._source[this._index + 1] !== "]") {
-      throw this.error('Expected "]]" to close a section');
-    }
-
-    this._index += 2;
   }
 
   /**
@@ -127,6 +99,8 @@ class RuntimeParser {
 
     if (this._source[this._index] === "=") {
       this._index++;
+    } else {
+      throw this.error("Expected \"=\" after the identifier");
     }
 
     this.skipInlineWS();
@@ -216,7 +190,7 @@ class RuntimeParser {
    * Get identifier using the provided regex.
    *
    * By default this will get identifiers of public messages, attributes and
-   * external arguments (without the $).
+   * variables (without the $).
    *
    * @returns {String}
    * @private
@@ -291,21 +265,30 @@ class RuntimeParser {
    * @private
    */
   getString() {
-    const start = this._index + 1;
+    let value = "";
+    this._index++;
 
-    while (++this._index < this._length) {
+    while (this._index < this._length) {
       const ch = this._source[this._index];
 
       if (ch === '"') {
+        this._index++;
         break;
       }
 
       if (ch === "\n") {
         throw this.error("Unterminated string expression");
       }
+
+      if (ch === "\\") {
+        value += this.getEscapedCharacter(["{", "\\", "\""]);
+      } else {
+        this._index++;
+        value += ch;
+      }
     }
 
-    return this._source.substring(start, this._index++);
+    return value;
   }
 
   /**
@@ -329,10 +312,17 @@ class RuntimeParser {
       eol = this._length;
     }
 
-    const firstLineContent = start !== eol ?
-      this._source.slice(start, eol) : null;
+    // If there's any text between the = and the EOL, store it for now. The next
+    // non-empty line will decide what to do with it.
+    const firstLineContent = start !== eol
+      // Trim the trailing whitespace in case this is a single-line pattern.
+      // Multiline patterns are parsed anew by getComplexPattern.
+      ? this._source.slice(start, eol).replace(trailingWSRe, "")
+      : null;
 
-    if (firstLineContent && firstLineContent.includes("{")) {
+    if (firstLineContent
+      && (firstLineContent.includes("{")
+        || firstLineContent.includes("\\"))) {
       return this.getComplexPattern();
     }
 
@@ -419,13 +409,19 @@ class RuntimeParser {
         }
         ch = this._source[this._index];
         continue;
-      } else if (ch === "\\") {
-        const ch2 = this._source[this._index + 1];
-        if (ch2 === '"' || ch2 === "{" || ch2 === "\\") {
-          ch = ch2;
-          this._index++;
-        }
-      } else if (ch === "{") {
+      }
+
+      if (ch === undefined) {
+        break;
+      }
+
+      if (ch === "\\") {
+        buffer += this.getEscapedCharacter();
+        ch = this._source[this._index];
+        continue;
+      }
+
+      if (ch === "{") {
         // Push the buffer to content array right before placeable
         if (buffer.length) {
           content.push(buffer);
@@ -437,18 +433,13 @@ class RuntimeParser {
         buffer = "";
         content.push(this.getPlaceable());
 
-        this._index++;
-
-        ch = this._source[this._index];
+        ch = this._source[++this._index];
         placeables++;
         continue;
       }
 
-      if (ch) {
-        buffer += ch;
-      }
-      this._index++;
-      ch = this._source[this._index];
+      buffer += ch;
+      ch = this._source[++this._index];
     }
 
     if (content.length === 0) {
@@ -456,12 +447,41 @@ class RuntimeParser {
     }
 
     if (buffer.length) {
-      content.push(buffer);
+      // Trim trailing whitespace, too.
+      content.push(buffer.replace(trailingWSRe, ""));
     }
 
     return content;
   }
   /* eslint-enable complexity */
+
+  /**
+   * Parse an escape sequence and return the unescaped character.
+   *
+   * @returns {string}
+   * @private
+   */
+  getEscapedCharacter(specials = ["{", "\\"]) {
+    this._index++;
+    const next = this._source[this._index];
+
+    if (specials.includes(next)) {
+      this._index++;
+      return next;
+    }
+
+    if (next === "u") {
+      const sequence = this._source.slice(this._index + 1, this._index + 5);
+      if (unicodeEscapeRe.test(sequence)) {
+        this._index += 5;
+        return String.fromCodePoint(parseInt(sequence, 16));
+      }
+
+      throw this.error(`Invalid Unicode escape sequence: \\u${sequence}`);
+    }
+
+    throw this.error(`Unknown escape sequence: \\${next}`);
+  }
 
   /**
    * Parses a single placeable in a Message pattern and returns its
@@ -499,7 +519,7 @@ class RuntimeParser {
     const ch = this._source[this._index];
 
     if (ch === "}") {
-      if (selector.type === "attr" && selector.id.name.startsWith("-")) {
+      if (selector.type === "getattr" && selector.id.name.startsWith("-")) {
         throw this.error(
           "Attributes of private messages cannot be interpolated."
         );
@@ -516,11 +536,11 @@ class RuntimeParser {
       throw this.error("Message references cannot be used as selectors.");
     }
 
-    if (selector.type === "var") {
+    if (selector.type === "getvar") {
       throw this.error("Variants cannot be used as selectors.");
     }
 
-    if (selector.type === "attr" && !selector.id.name.startsWith("-")) {
+    if (selector.type === "getattr" && !selector.id.name.startsWith("-")) {
       throw this.error(
         "Attributes of public messages cannot be used as selectors."
       );
@@ -558,6 +578,10 @@ class RuntimeParser {
    * @private
    */
   getSelectorExpression() {
+    if (this._source[this._index] === "{") {
+      return this.getPlaceable();
+    }
+
     const literal = this.getLiteral();
 
     if (literal.type !== "ref") {
@@ -570,7 +594,7 @@ class RuntimeParser {
       const name = this.getIdentifier();
       this._index++;
       return {
-        type: "attr",
+        type: "getattr",
         id: literal,
         name
       };
@@ -582,7 +606,7 @@ class RuntimeParser {
       const key = this.getVariantKey();
       this._index++;
       return {
-        type: "var",
+        type: "getvar",
         id: literal,
         key
       };
@@ -620,7 +644,7 @@ class RuntimeParser {
     const args = [];
 
     while (this._index < this._length) {
-      this.skipInlineWS();
+      this.skipWS();
 
       if (this._source[this._index] === ")") {
         return args;
@@ -637,7 +661,7 @@ class RuntimeParser {
 
         if (this._source[this._index] === ":") {
           this._index++;
-          this.skipInlineWS();
+          this.skipWS();
 
           const val = this.getSelectorExpression();
 
@@ -665,7 +689,7 @@ class RuntimeParser {
         }
       }
 
-      this.skipInlineWS();
+      this.skipWS();
 
       if (this._source[this._index] === ")") {
         break;
@@ -865,7 +889,7 @@ class RuntimeParser {
     if (cc0 === 36) { // $
       this._index++;
       return {
-        type: "ext",
+        type: "var",
         name: this.getIdentifier()
       };
     }
@@ -905,12 +929,11 @@ class RuntimeParser {
     // to parse them properly and skip their content.
     let eol = this._source.indexOf("\n", this._index);
 
-    while (eol !== -1 &&
-      ((this._source[eol + 1] === "/" && this._source[eol + 2] === "/") ||
-       (this._source[eol + 1] === "#" &&
-         [" ", "#"].includes(this._source[eol + 2])))) {
-      this._index = eol + 3;
+    while (eol !== -1
+      && this._source[eol + 1] === "#"
+      && [" ", "#"].includes(this._source[eol + 2])) {
 
+      this._index = eol + 3;
       eol = this._source.indexOf("\n", this._index);
 
       if (eol === -1) {
@@ -952,7 +975,7 @@ class RuntimeParser {
 
         if ((cc >= 97 && cc <= 122) || // a-z
             (cc >= 65 && cc <= 90) || // A-Z
-             cc === 47 || cc === 91) { // /[
+             cc === 45) { // -
           this._index = start;
           return;
         }